feat: extract aggregation logic to unit module
