+++ src/modules/order/order.module.ts
@@ -4,11 +4,10 @@ import { OrderController } from './order.controller';
 import { PrismaModule } from '../../prisma/prisma.module';
 import { SellerApiModule } from '@/api/seller/seller.module';
 import { OrderRepository } from './order.repository';
-import { TransactionRepository } from '@/modules/transaction/transaction.repository';
 
 @Module({
   imports: [PrismaModule, SellerApiModule],
   controllers: [OrderController],
-  providers: [OrderService, OrderRepository, TransactionRepository],
+  providers: [OrderService, OrderRepository],
 })
 export class OrderModule {}
